+++ impl/no-deps-transformer/src/main/java/org/wildfly/transformer/nodeps/ClassFileUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.javax2jakarta;
+package org.wildfly.transformer.nodeps;
 
 /**
  * Utility class for working with class file content.
